Fix linter issue
